requirements.txt
playwright
beautifulsoup4
psycopg2-binary
python-dotenv